PlayConnected: component now uses model Nodes and improves lexing.
In order to simplify API fetching, the Redux store now holds current
graph data in the same structure as it is stored in the database. This
means we now have to map the structures to ones ResponsiveNetwork can
use from within the component.

We also improved the lexing stage when submitting a new response,
stripping special characters, collapsing whitespace, and splitting
tokents on spaces and hyphens.