Explore: show real data for user history of current calendar year.
We now fetch actual user history from the current calendar year to
display in ResponsiveCalendar. Sessions are grouped by day, memoized,
and displayed in the heatmap. Because the business logic for
calculating dates is clunky, we should eventually refactor to use more
elegant solution from date-fns package to accomplish this.

We also calculate some usage statistics like total words logged and
week-over-week activity. It would be great to do this on the backend
ultimately, but for now this works!

Days are now clickable to expose session data for that date. The next
step is to add a log page that will show the relevant sessions and
allow further exploration.