graphActions: extract all parsing and tree-shaking into util functions
We used to parse the input string and shake the node/link trees from
within the Play component. Now these functions are separated into their
own utilities and are called from a single submitResponse action.

In future, we will add the complete response string to the database as
well, and can take a similar approach.