Play: add loading skeletons, resumability and finishability to sessions
Before initializing a session, we prompt the user to either create a new
one, resume the previous, or cancel to return to the last page. This is
implemented with a nice full-screen confirm.

We also add the option to finish a running session, giving the user the
option to close out their work and start another, or to resume it at
another time.