graphActions: fix linking of existing edges
We previously forgot to add new edges to existing nodes to the backend.
This caused incorrect and sparse trees in the session viewer. We now
have a more robust (if verbose) method of iterating existing nodes to
find the correct ones to link.

There is now a new set of starting words, which we have moved along with
other constants to src/data/constants.js.