graphActions: remove stale nodes from unplayed sessions
When a user starts a session and promptly ends it without adding any
words, we now delete the initial node from the DB to keep the unused
data to a minimum.

We still need to remove the actual WordNet itself, but due to a race
condition (the node gets deleted after it's parent network) this throws
an error *sometimes*. May look into a custom resolver for this.