SessionLog: fix timezone discrepancy when querying by createdAt.
When a date string is passed as a URL param in the form 'yyyy-MM-dd', we
now parse this into a local date object, which we use to find start and
end timestamps in unix time for the local calendar date. This can easily
be filtered on when retreiving sessions.

Next, we will build a date picker component to make date and date-range
selection easier.