SessionLog: improved date filtering and breadcrumb navigation
The initial date filter is now provided by the Explore component.
When a calendar day is clicked, we use the date string to generate a
startTimestamp and endTimestamp representing the local start and end of
the day in Unix Time, via date-fns. If no initialDate is provided, we
use an arbitrary very-small and very-large time. These are then used as
the filter parameters for the query.

Existing date filter is shown as a closeable tag above the table, which
upon closing will clear the start and end dates.

In doing so, we moved the date out of a URL parameter and into component
state, which is a more pleasant pattern. This means we no longer see the
date as a crumb in the breadcrumnb.