Explore: add dowload option and improve upon Session views
We now have a Download component which takes an optional 'graph' object
or 'all' prop, and exports either that graph data, or all user graph
data to an Excel .xlsx file. This appears in the SessionLog with the
default 'all' behavior, and in a Session view with the default 'graph'
behavior. The component renders as whatever 'element' prop is passed to
it, and currently manages its own loading state based on the status of
the graphql operation, if any. This mean that any component that
implements the 'loading' prop will use its default loading appearance
while fetching.

Session now shows four tabs in the lower section, consisting of:
'Overview' which shows general info for the graph, including the node,
edge, and response count, and the graph density; 'Nodes' which now also
shows the degree of each node; 'Edges' which is more or less unchanged;
and 'Responses' which shows each response and the time taken.