Download: delay async load of session data until triggered
We added a 'render' prop which accepts a function which takes an object
with a 'loading' and 'onClick' property, and returns a component to
serve as the display component.

Data is now not fetched unless the component is triggered, which
improves the experience on the SessionLog page greatly. Unfortunately,
we are still seeing performance issues on the page AFTER the
load/download has completed. Will pursue this at a later time.