Explore: refactor Explore route to reduce coupling and network calls
We now use a simple stateless router for the root of the explore routes,
bypassing the coupling of the dashboard and the SessionLog. The deisplay
logic was moved into a separate Dashboard component, which communicates
the date parameters on calendar click to the SessionLog via the redux
store.

An added benefit of this is that we do not have to perform the dashboard
fetching if we go straight to SessionLog or to a Session, such as after
completing a game.