graphUtils: extend useGraph hook with a useTraversableGraph option
This hook extends the functionality of useGraph by adding two utility
functions to the return value -- dfs() and shortestPath().

The dfs() function takes a source string argument and performs a
depth-first search of the local subraph and (when implemented) returns
the eccentricity (maximum distance to any of its neighbors).

The shortestPath() function takes a source and target string, and
performs a Dijkstra search for the shortest path between them, returning
an array of the node values in that path, or null if no path found.