graphUtils: move useDensity() here and add useDiameter()
We relocated the useDensity hook to its appropriate place in the
graphUtils file, and added a useDiameter hook, which takes a mapped
graph and a bfs() function as parameters, and returns the largest
geodesic of the graph (its diameter). This is memoized and should only
be calculated once on graph load, due to its cost.

Ultimately we should rethink the design of this so we don't have to pass
in the bfs function.