ErrorBoundary: add new error boundary component to wrap temperamentals
This component follows React's suggested implementation for Error
Boundaries, allowing you to pass an optional 'fallback' prop to
render something on error. In future, maybe this could be a render
prop to allow customization of the error message shown.

We also fire an event to the Analytics service, reporting the error
message and location.