Dashboard: use vega-scale color scheme for calendar view
We are simply using vega to generate a static palette for the
heatmap. We tried using the existing 'colorScheme' to generate a
dynamic palette, but some of the results were not appropriate.

In future we can move the interpolator elsewhere so we don't have
to generate a new palette on each mount.