historyActions: correct fetchAllSessions() to now exhaustively fetch
We do a proper chaining of queries until all WordNet records are
fetched. In addition, we condense all graphs into a single graph, with
the help of condenseDomainNodes() and condenseDomainEdges() functions in
graphUtils.

This is tested and working, but definitely has room for improvement.
Performance issues may largely lie with the graph view itself, but there
are certainly opporunities.