GraphQL: implement BatchPut for Nodes and Edges
This patch adds two GraphQL mutations and their associated resolvers to
aid in adding graph components. On response submission, instead of
chaining individual mutations (one for each new node, then one for each
edge based on the returned IDs), we now have simple batchCreateNodes()
and batchCreateEdges() mutations which can accomplish the same in only
two round-trips.

In future, it would be nice to combine this advancement along with
createResponse() into a single, atomic mutation, and relocate the
response-parsing logic to a lambda function.