Download: refactor and fix 'download all' behavior
By plugging in to the Redux store, we have greatly improved readability
of this component, relying on the same logic for downloading one session
as for many. This includes mapping the graph components from domain
representation to moodel representation.

Doing so allowed us to remove the 'graph' prop, instead using the
'currentSession' from the store to populate the table.