Assessments: scaffold assessment flow and begin Grid assessment impl
This patch adds basic functionality for the PictographGridAssessment, in
which users place provided pictographs into a 4x4 grid to create a scene
as they see fit. Using react-dnd, we present a panel of pictographs
which can be dragged and dropped into a position of the grid. Each grid
square can hold one pictograph, and they can be repositioned or deleted
from the grid.

Next we will need to work on tying this in with the backend
implementation. The question is whether to create a profile model in the
database at the start of the assessment, and issue udpate mutations as
steps are completed, or to wait until the flow is completed to submit a
single create operation.