Assessment: initial implementation of ShapeDiscriminationAssessment
We now have a ShapeDiscriminationAssessment, in which the user is asked
to place shapes of differing type, size, and color, into any of three
groups as they see fit. The initial shapes are static, but their order
is randomized when the component is first mounted. The user is not
allowed to proceed to the next step until all shapes have been assigned
to one of the three groups.