Assessment: initial implementation of TimelineAssessment
We now have a TimelineAssessment component, which asks the user to order
a number of preselected pictographs in a timeline to tell a story or to
inform about themself. When dragging to reorder, we use the hovered and
dragged items' client rects to determine a hover, after which we insert
the dragged item at that point in the list and shift the following
elements.

It would be nice to detect the x-offset of the hover, so that hovering
on the right side of an item would place it after the item, and hovering
on the left side of the item would place it before.