Assessment: add remaining assessments and allow each step to navigate
We now delegate navigation button logic and rendering to each assesment,
allowing them to render their own "previous" and "next" buttons, and
determine when to allow navigation to another step. This can be used to
prevent a user from leaving an assessment in an incomplete state.

Though this is not a bad pattern, it would be nice to render the buttons
once instead of in each component, and allow them to override the
navigation logic...