Assessment: refactor to use redux to manage assessment states
In doing so, we will be able to have access to all assessment states at
the end of all steps, in order to generate a report and push data to the
server. If in the future we decide we would like assessment updates to
be pushed incrementally, we may want to return to using component-level
state only.

While this moves some fetching logic out of the components an into the
action creators, it adds all the usual boilerplate noise for connected
components.